Unifies params into one object.
